+++ src/nvim/indent.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "nvim/ex_cmds_defs.h"  // IWYU pragma: keep
+#include "nvim/normal_defs.h"
 #include "nvim/pos_defs.h"  // IWYU pragma: keep
 #include "nvim/types_defs.h"  // IWYU pragma: keep
 
@@ -13,6 +14,8 @@ enum {
   SIN_UNDO    = 4,  ///< save line for undo before changing it
   SIN_NOMARK  = 8,  ///< don't adjust extmarks
 };
+
+typedef int (*Indenter)(void);
 
 #ifdef INCLUDE_GENERATED_DECLARATIONS
 # include "indent.h.generated.h"
