+++ src/nvim/indent_c.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "nvim/eval/typval_defs.h"
 #include "nvim/pos_defs.h"  // IWYU pragma: keep
 #include "nvim/types_defs.h"  // IWYU pragma: keep
